added Get-FTPcommesse

+++ _others/Get-FTPcommesse.ps1
@@ -1,23 +1,63 @@
-# Remote Path (CASE SENSITIVE: triple quotes are required for the script to consider the spaces)
-# Purtroppo bisogna inserire il nome esatto della commessa
+<#
+.Synopsis
+    Automatically connects to Bitech's server and download VR (Released Version) from a specific order.
+.DESCRIPTION
+    The script uses WinSCP CLI to connect to Bitech's SFTP server and downloads the listed documents.
+    The $Customer variable contains the order's end customer.
+    The documents must match the VR_ (Released Version) prefix and can be .doc and .pdf.
+    The destination path is customizable via the $RootPath variable.
+    All the empty documents and folders are then removed.
+    The performance of the script is tracked via C# stopwatch class and logged.
+.EXAMPLE
+    PS> ./Get-FTPcommesse.ps1
+.NOTES
+    It requires WinSCP installed in C:\Program Files(x86)\WinSCP\WinSCP.exe
+    The order's name must be precise and case-sensitive.
+    Just edit $Customer and $RootPath.
+#>
 
-$RemotePath = '""/MC_Commesse/CO BCube""'
-$LocalPath = 'C:\_TEMP'
-$LocalLog = "$LocalPath\WinSCP_commesse.log"
+
+# Remote Path (CASE SENSITIVE)
+
+$Customer = 'RealeS' # Inserire il nome esatto della commessa!
+
+$RemotePath = '""/MC_Commesse/CO ' + $Customer + '""'   # triple quotes are required for the script to consider the spaces
+
+# RootPath: dove si vogliono salvare i documenti finali; viene anche creato un log
+
+$RootPath = 'C:\_TEMP'
+$LocalLog = "$RootPath\WinSCP_commesse.log"
+
+# LocalPath (CASE SENSITIVE) : cartella temporanea in cui verranno salvati tutti i documenti prima di essere filtrati
+
+New-Item -Path . -Name "temp" -ItemType "Directory"
+$LocalPath = "$RootPath\temp"
 
 # Open connection and execute commands: execution is timed
 
 $Clock = [Diagnostics.Stopwatch]::StartNew()
 
-Remove-Item -Path "$LocalLog"
+if ( Test-Path "$RootPath\WinSCP_commesse.log" ) {
+  Remove-Item -Path "$LocalLog"
+}
 
 & "C:\Program Files (x86)\WinSCP\WinSCP.com" `
-  /log="C:\_TEMP\WinSCP_commesse.log" /ini=nul `
+  /log="$LocalLog" /ini=nul `
   /command `
     "open ftpes://sanna:edo89%2B0304@79.11.21.211/ -certificate=`"`"3f:3f:9f:7a:49:0e:4d:80:12:69:af:70:cb:5c:72:a4:e7:3a:eb:f1`"`"" `
     "cd $RemotePath" `
-    "get -filemask=VR_*.doc $RemotePath $LocalPath"
-    # TODO: da completare, adesso esporta anche le cartelle vuote
+    "get -filemask=VR_*.doc $RemotePath $LocalPath" `
+    "exit"
+
+# Spostare i file ricavati da LocalPath a RootPath, rimuovendo tutte le cartelle vuote e le commesse non chiuse
+
+Foreach ( $file in $(Get-ChildItem -Recurse *.doc, *.pdf) ) {
+  Move-Item -Path $file -Destination $RootPath
+}
+Remove-Item -Recurse $LocalPath
+if ( Test-Path -Path "$RootPath\VR_Versione_rilasciata.doc" ) {
+  Remove-Item -Path "$RootPath\VR_Versione_rilasciata.doc"
+}
 
 # Exit code
 
@@ -31,4 +71,4 @@ if ($winscpResult -eq 0) {
 # Track performance
 
 $Clock.Stop()
-Add-Content -Path "$LocalLog" -Value "Execution time approx. $($Clock.Elapsed.Minutes) minutes"
+Add-Content -Path "$LocalLog" -Value "Execution time approx. $($Clock.Elapsed.Minutes) minutes $($Clock.Elapsed.Seconds) seconds"
